pyproject.toml
[tool.poetry]
name = "django-file-processing"
version = "0.1.0"
description = ""
authors = ["Nyaaa <1457047+Nyaaa@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.1"
djangorestframework = "^3.14.0"
celery = {version = "^5.3.6", extras = ["redis"]}

[tool.poetry.group.dev.dependencies]
coverage = "^7.4.1"
drf-spectacular = "^0.27.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
branch = true
source = ['.']
omit = ['*/tests/*', '*/migrations/*', 'file_uploader/*']

[tool.coverage.report]
skip_covered = true

[tool.ruff]
line-length = 119
exclude = ['*/migrations/*']
select = ['E', 'F', 'UP', 'NPY', 'SIM', 'ICN', 'C4', 'A', 'B', 'N', 'I', 'Q', 'DJ',
    'W', 'PL', 'RET', 'RUF', 'PERF', 'PIE', 'ARG', 'PTH', 'PD', 'PGH', 'TRY']
ignore = ['TRY003', 'ARG002', 'RUF012']

[tool.ruff.flake8-quotes]
inline-quotes = 'single'
multiline-quotes = 'double'
